Moving threadpool libraries to tensorflow/core/platform.
PiperOrigin-RevId: 269123124
