+++ tensorflow/core/platform/threadpool.cc
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow/core/lib/core/threadpool.h"
+#include "tensorflow/core/platform/threadpool.h"
 
 #define EIGEN_USE_THREADS
